% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fteTheme.R
\name{fte_theme}
\alias{fte_theme}
\title{fte_theme}
\usage{
fte_theme()
}
\description{
Provides a theme similar to the Five Thirty Eight Blog.  Code found from: http://minimaxir.com/2015/02/ggplot-tutorial/ created by Max Woolf
}

